scripts/requirements.txt
azure-search-documents>=11.4.0
azure-core>=1.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
